address tom's comments
