+++ prbench-imitation-learning/scripts/train_lerobot_direct.py
@@ -1,358 +1,12 @@
 #!/usr/bin/env python
-"""Direct LeRobot training script for PushT.
+"""Direct LeRobot training script.
 
-This script mimics the official lerobot_train.py and provides a streamlined
-training pipeline using the LeRobot framework directly.
-
-Usage:
-    python scripts/train_lerobot_direct.py --output_dir results/pusht_run1 --steps 200000
-    python scripts/train_lerobot_direct.py --output_dir results/pusht_run1 \
-    --steps 200000 --eval_freq 10000
+for both lerobot and PRBench environments.
 """
 
-import logging
-import sys
-import time
-from contextlib import nullcontext
-from pprint import pformat
+from lerobot.utils.utils import init_logging
 
-import torch
-from lerobot.configs import parser
-from lerobot.configs.train import TrainPipelineConfig
-from lerobot.datasets.factory import make_dataset
-from lerobot.datasets.sampler import EpisodeAwareSampler
-from lerobot.datasets.utils import cycle
-from lerobot.envs.factory import make_env, make_env_config
-from lerobot.envs.utils import close_envs
-from lerobot.optim.factory import make_optimizer_and_scheduler
-from lerobot.policies.factory import make_policy, make_pre_post_processors
-from lerobot.rl.wandb_utils import WandBLogger
-from lerobot.utils.logging_utils import AverageMeter, MetricsTracker
-from lerobot.utils.random_utils import set_seed
-from lerobot.utils.train_utils import (
-    get_step_checkpoint_dir,
-    get_step_identifier,
-    load_training_state,
-    save_checkpoint,
-    update_last_checkpoint,
-)
-from lerobot.utils.utils import (
-    format_big_number,
-    get_safe_torch_device,
-    init_logging,
-)
-from lerobot_eval import eval_policy_all
-from termcolor import colored
-from torch.amp import GradScaler
-
-from prbench_imitation_learning.train import update_policy
-
-
-@parser.wrap()
-def train(cfg: TrainPipelineConfig):
-    """Main function to train a policy using LeRobot framework.
-
-    This function orchestrates the entire training pipeline, including:
-    - Setting up logging, seeding, and device configuration.
-    - Creating the dataset, evaluation environment, policy, and optimizer.
-    - Handling resumption from a checkpoint.
-    - Running the main training loop, which involves
-    - fetching data batches and calling `update_policy`.
-    - Periodically logging metrics, saving model checkpoints, and evaluating the policy.
-    - Pushing the final trained model to the Hugging Face Hub if configured.
-
-    Args:
-        cfg: A `TrainPipelineConfig` object containing all training configurations.
-    """
-    cfg.validate()
-    logging.info(pformat(cfg.to_dict()))
-
-    if cfg.wandb.enable and cfg.wandb.project:
-        wandb_logger = WandBLogger(cfg)
-    else:
-        wandb_logger = None
-        logging.info(colored("Logs will be saved locally.", "yellow", attrs=["bold"]))
-
-    if cfg.seed is not None:
-        set_seed(cfg.seed)
-
-    # Check device is available
-    device = get_safe_torch_device(cfg.policy.device, log=True)
-    torch.backends.cudnn.benchmark = True
-    torch.backends.cuda.matmul.allow_tf32 = True
-
-    logging.info("Creating dataset")
-    dataset = make_dataset(cfg)
-
-    # Create environment used for evaluating checkpoints during training
-    # on simulation data.
-    eval_env = None
-    eval_env_cfg = cfg.env
-
-    # CLI override: --eval_env.type=... --eval_env.task=...
-    eval_env_type = None
-    eval_env_task = None
-    for arg in sys.argv:
-        if arg.startswith("--eval_env.type="):
-            eval_env_type = arg.split("=", 1)[1]
-        elif arg.startswith("--eval_env.task="):
-            eval_env_task = arg.split("=", 1)[1]
-
-    if cfg.eval_freq > 0:
-        if eval_env_type is not None:
-            logging.info(
-                f"Using eval env from CLI: type={eval_env_type}, task={eval_env_task}"
-            )
-            eval_env_cfg = make_env_config(eval_env_type, task=eval_env_task)
-        elif eval_env_cfg is None:
-            # If user didn't pass an env, auto-create a sensible default from dataset
-            try:
-                robot_type = dataset.meta.robot_type or ""
-            except Exception:
-                robot_type = ""
-            if isinstance(robot_type, str) and "motion2d" in robot_type.lower():
-                logging.info(
-                    "No --env specified;"
-                    "auto-configuring prbench Motion2D-p0-v0 for evaluation"
-                )
-                eval_env_cfg = make_env_config("prbench", task="Motion2D-p0-v0")
-    if cfg.eval_freq > 0 and eval_env_cfg is not None:
-        logging.info("Creating env")
-        eval_env = make_env(
-            eval_env_cfg,
-            n_envs=cfg.eval.batch_size,
-            use_async_envs=cfg.eval.use_async_envs,
-        )
-
-    logging.info("Creating policy")
-    policy = make_policy(
-        cfg=cfg.policy,
-        ds_meta=dataset.meta,
-    )
-
-    # Create processors - only provide dataset_stats
-    # if not resuming from saved processors
-    processor_kwargs = {}
-    postprocessor_kwargs = {}
-    if (
-        cfg.policy.pretrained_path and not cfg.resume
-    ) or not cfg.policy.pretrained_path:
-        # Only provide dataset_stats when not resuming from saved processor state
-        processor_kwargs["dataset_stats"] = dataset.meta.stats
-
-    if cfg.policy.pretrained_path is not None:
-        processor_kwargs["preprocessor_overrides"] = {
-            "device_processor": {"device": device.type},
-            "normalizer_processor": {
-                "stats": dataset.meta.stats,
-                "features": {
-                    **policy.config.input_features,
-                    **policy.config.output_features,
-                },
-                "norm_map": policy.config.normalization_mapping,
-            },
-        }
-        postprocessor_kwargs["postprocessor_overrides"] = {
-            "unnormalizer_processor": {
-                "stats": dataset.meta.stats,
-                "features": policy.config.output_features,
-                "norm_map": policy.config.normalization_mapping,
-            },
-        }
-
-    preprocessor, postprocessor = make_pre_post_processors(
-        policy_cfg=cfg.policy,
-        pretrained_path=cfg.policy.pretrained_path,
-        **processor_kwargs,
-        **postprocessor_kwargs,
-    )
-
-    logging.info("Creating optimizer and scheduler")
-    optimizer, lr_scheduler = make_optimizer_and_scheduler(cfg, policy)
-    grad_scaler = GradScaler(device.type, enabled=cfg.policy.use_amp)
-
-    step = 0  # number of policy updates (forward + backward + optim)
-
-    if cfg.resume:
-        step, optimizer, lr_scheduler = load_training_state(
-            cfg.checkpoint_path, optimizer, lr_scheduler
-        )
-
-    num_learnable_params = sum(
-        p.numel() for p in policy.parameters() if p.requires_grad
-    )
-    num_total_params = sum(p.numel() for p in policy.parameters())
-
-    logging.info(  # pylint: disable=logging-not-lazy
-        colored("Output dir:", "yellow", attrs=["bold"]) + f" {cfg.output_dir}"
-    )
-    if eval_env is not None:
-        try:
-            logging.info(f"{eval_env_cfg.task=}")
-        except Exception:
-            pass
-    logging.info(f"{cfg.steps=} ({format_big_number(cfg.steps)})")
-    logging.info(f"{dataset.num_frames=} ({format_big_number(dataset.num_frames)})")
-    logging.info(f"{dataset.num_episodes=}")
-    logging.info(f"{num_learnable_params=} ({format_big_number(num_learnable_params)})")
-    logging.info(f"{num_total_params=} ({format_big_number(num_total_params)})")
-
-    # create dataloader for offline training
-    if hasattr(cfg.policy, "drop_n_last_frames"):
-        shuffle = False
-        sampler = EpisodeAwareSampler(
-            dataset.meta.episodes["dataset_from_index"],
-            dataset.meta.episodes["dataset_to_index"],
-            drop_n_last_frames=cfg.policy.drop_n_last_frames,
-            shuffle=True,
-        )
-    else:
-        shuffle = True
-        sampler = None
-
-    dataloader = torch.utils.data.DataLoader(
-        dataset,
-        num_workers=cfg.num_workers,
-        batch_size=cfg.batch_size,
-        shuffle=shuffle and not cfg.dataset.streaming,
-        sampler=sampler,
-        pin_memory=device.type == "cuda",
-        drop_last=False,
-        prefetch_factor=2,
-    )
-    dl_iter = cycle(dataloader)
-
-    policy.train()
-
-    train_metrics = {
-        "loss": AverageMeter("loss", ":.3f"),
-        "grad_norm": AverageMeter("grdn", ":.3f"),
-        "lr": AverageMeter("lr", ":0.1e"),
-        "update_s": AverageMeter("updt_s", ":.3f"),
-        "dataloading_s": AverageMeter("data_s", ":.3f"),
-    }
-
-    train_tracker = MetricsTracker(
-        cfg.batch_size,
-        dataset.num_frames,
-        dataset.num_episodes,
-        train_metrics,
-        initial_step=step,
-    )
-
-    logging.info("Start offline training on a fixed dataset")
-    for _ in range(step, cfg.steps):
-        start_time = time.perf_counter()
-        batch = next(dl_iter)
-        batch = preprocessor(batch)
-        train_tracker.dataloading_s = time.perf_counter() - start_time
-
-        train_tracker, output_dict = update_policy(
-            train_tracker,
-            policy,
-            batch,
-            optimizer,
-            cfg.optimizer.grad_clip_norm,
-            grad_scaler=grad_scaler,
-            lr_scheduler=lr_scheduler,
-            use_amp=cfg.policy.use_amp,
-        )
-
-        # Note: eval and checkpoint happens *after* the
-        # `step`th training update has completed, so we
-        # increment `step` here.
-        step += 1
-        train_tracker.step()
-        is_log_step = cfg.log_freq > 0 and step % cfg.log_freq == 0
-        is_saving_step = step % cfg.save_freq == 0 or step == cfg.steps
-        is_eval_step = cfg.eval_freq > 0 and step % cfg.eval_freq == 0
-
-        if is_log_step:
-            logging.info(train_tracker)
-            if wandb_logger:
-                wandb_log_dict = train_tracker.to_dict()
-                if output_dict:
-                    wandb_log_dict.update(output_dict)
-                wandb_logger.log_dict(wandb_log_dict, step)
-            train_tracker.reset_averages()
-
-        if cfg.save_checkpoint and is_saving_step:
-            logging.info(f"Checkpoint policy after step {step}")
-            checkpoint_dir = get_step_checkpoint_dir(cfg.output_dir, cfg.steps, step)
-            save_checkpoint(
-                checkpoint_dir,
-                step,
-                cfg,
-                policy,
-                optimizer,
-                lr_scheduler,
-                preprocessor,
-                postprocessor,
-            )
-            update_last_checkpoint(checkpoint_dir)
-            if wandb_logger:
-                wandb_logger.log_policy(checkpoint_dir)
-
-        if cfg.env and is_eval_step:
-            step_id = get_step_identifier(step, cfg.steps)
-            logging.info(f"Eval policy at step {step}")
-            with (
-                torch.no_grad(),
-                (
-                    torch.autocast(device_type=device.type)
-                    if cfg.policy.use_amp
-                    else nullcontext()
-                ),
-            ):
-                eval_info = eval_policy_all(
-                    envs=eval_env,  # type: ignore
-                    policy=policy,
-                    preprocessor=preprocessor,
-                    postprocessor=postprocessor,
-                    n_episodes=cfg.eval.n_episodes,
-                    videos_dir=cfg.output_dir / "eval" / f"videos_step_{step_id}",
-                    max_episodes_rendered=4,
-                    start_seed=cfg.seed,
-                    max_parallel_tasks=cfg.env.max_parallel_tasks,
-                )
-            # overall metrics (suite-agnostic)
-            aggregated = eval_info["overall"]
-
-            # optional: per-suite logging
-            for suite, suite_info in eval_info.items():
-                logging.info("Suite %s aggregated: %s", suite, suite_info)
-
-            # meters/tracker
-            eval_metrics = {
-                "avg_sum_reward": AverageMeter("∑rwrd", ":.3f"),
-                "pc_success": AverageMeter("success", ":.1f"),
-                "eval_s": AverageMeter("eval_s", ":.3f"),
-            }
-            eval_tracker = MetricsTracker(
-                cfg.batch_size,
-                dataset.num_frames,
-                dataset.num_episodes,
-                eval_metrics,
-                initial_step=step,
-            )
-            eval_tracker.eval_s = aggregated.pop("eval_s")
-            eval_tracker.avg_sum_reward = aggregated.pop("avg_sum_reward")
-            eval_tracker.pc_success = aggregated.pop("pc_success")
-            if wandb_logger:
-                wandb_log_dict = {**eval_tracker.to_dict(), **eval_info}
-                wandb_logger.log_dict(wandb_log_dict, step, mode="eval")
-                wandb_logger.log_video(
-                    eval_info["overall"]["video_paths"][0], step, mode="eval"
-                )
-
-    if eval_env:
-        close_envs(eval_env)
-    logging.info("End of training")
-
-    if cfg.policy.push_to_hub:
-        policy.push_model_to_hub(cfg)
-        preprocessor.push_to_hub(cfg.policy.repo_id)
-        postprocessor.push_to_hub(cfg.policy.repo_id)
+from prbench_imitation_learning.train import train
 
 
 def main() -> None:
